requirements_aws.txt
# requirements.txt actualizado para AWS
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
langchain==0.0.350
faiss-cpu==1.7.4
openai==1.3.7
redis==5.0.1
boto3==1.34.0
botocore==1.34.0
pillow==10.1.0
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
websockets==12.0
jinja2==3.1.2
aiofiles==23.2.1









